move refineExpression, dont pass parser around flattenExpression unnecessarily
