+++ src/parse/utils/flattenExpression.js
@@ -8,7 +8,7 @@ export default function flattenExpression ( expression ) {
 
 	flattened = {
 		r: refs,
-		s: stringify( this, expression, refs )
+		s: stringify( expression, refs )
 	};
 
 	return flattened;
@@ -53,9 +53,9 @@ function extractRefs ( node, refs ) {
 	}
 }
 
-function stringify ( parser, node, refs ) {
+function stringify ( node, refs ) {
 	var stringifyAll = function ( item ) {
-		return stringify( parser, item, refs );
+		return stringify( item, refs );
 	};
 
 	switch ( node.t ) {
@@ -74,33 +74,33 @@ function stringify ( parser, node, refs ) {
 		return '{' + ( node.m ? node.m.map( stringifyAll ).join( ',' ) : '' ) + '}';
 
 		case KEY_VALUE_PAIR:
-		return node.k + ':' + stringify( parser, node.v, refs );
+		return node.k + ':' + stringify( node.v, refs );
 
 		case PREFIX_OPERATOR:
-		return ( node.s === 'typeof' ? 'typeof ' : node.s ) + stringify( parser, node.o, refs );
+		return ( node.s === 'typeof' ? 'typeof ' : node.s ) + stringify( node.o, refs );
 
 		case INFIX_OPERATOR:
-		return stringify( parser, node.o[0], refs ) + ( node.s.substr( 0, 2 ) === 'in' ? ' ' + node.s + ' ' : node.s ) + stringify( parser, node.o[1], refs );
+		return stringify( node.o[0], refs ) + ( node.s.substr( 0, 2 ) === 'in' ? ' ' + node.s + ' ' : node.s ) + stringify( node.o[1], refs );
 
 		case INVOCATION:
-		return stringify( parser, node.x, refs ) + '(' + ( node.o ? node.o.map( stringifyAll ).join( ',' ) : '' ) + ')';
+		return stringify( node.x, refs ) + '(' + ( node.o ? node.o.map( stringifyAll ).join( ',' ) : '' ) + ')';
 
 		case BRACKETED:
-		return '(' + stringify( parser, node.x, refs ) + ')';
+		return '(' + stringify( node.x, refs ) + ')';
 
 		case MEMBER:
-		return stringify( parser, node.x, refs ) + stringify( parser, node.r, refs );
+		return stringify( node.x, refs ) + stringify( node.r, refs );
 
 		case REFINEMENT:
-		return ( node.n ? '.' + node.n : '[' + stringify( parser, node.x, refs ) + ']' );
+		return ( node.n ? '.' + node.n : '[' + stringify( node.x, refs ) + ']' );
 
 		case CONDITIONAL:
-		return stringify( parser, node.o[0], refs ) + '?' + stringify( parser, node.o[1], refs ) + ':' + stringify( parser, node.o[2], refs );
+		return stringify( node.o[0], refs ) + '?' + stringify( node.o[1], refs ) + ':' + stringify( node.o[2], refs );
 
 		case REFERENCE:
 		return '_' + refs.indexOf( node.n );
 
 		default:
-		parser.error( 'Expected legal JavaScript' );
+		throw new Error( 'Expected legal JavaScript' );
 	}
 }
